fix: fix routing to admin dashboard to always be protected
fix routing of admin dashboard to always be protected against unwanted access by adding layout check and moving some pages
