+++ frontend/app/admin/(protected)/dashboard/page.tsx
@@ -1,0 +1,5 @@
+import AdminContent from '@/src/components/admin/AdminContent';
+
+export default async function AdminDashboard() {
+  return <AdminContent />;
+}
