+++ frontend/app/admin/(protected)/suggestions/[id]/edit/page.tsx
