+++ frontend/app/admin/(protected)/suggestions/[id]/page.tsx
